+++ client/src/styles/Config/Config.css
@@ -35,31 +35,11 @@
   background-color: #25a351;
 }
 
-.Config .CurrentNumbers {
-  padding-top: 3em;
-  font-size: 3em;
-}
-
-.Config .AddNumber {
-  padding-top: 3em;
-  margin-bottom: 12em;
-}
-
-.Config .AddNumber input {
-  padding: 0.5em;
-  font-size: 1em;
-  border: none;
-}
-
-.Config .AddNumber input:hover {
-  border: 1px solid #1db954;
-}
-
-.Config .AddNumber button {
+.Config button {
   color: white;
   background-color: #1db954;
   border: none;
-  border-radius: .5em;
+  border-radius: 0.5em;
   padding: 1em;
   margin-left: 1em;
 }
