+++ client/src/styles/ExplorePage/ExplorePage.css
@@ -1,0 +1,15 @@
+.explorePage {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  padding: 10% 20%;
+  height: 100%;
+}
+/*# sourceMappingURL=ExplorePage.css.map */
