+++ package.json
@@ -26,7 +26,7 @@
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "js-cookie": "^2.2.1",
-    "mongoose": "^5.10.12",
+    "mongoose": "^5.11.4",
     "mongoose-simple-random": "^0.4.1",
     "morgan": "^1.10.0",
     "passport": "^0.4.1",
